adds integration test and moves to context
