+++ EloquaClient.Tests/Integration/ClientTests.cs
@@ -1,10 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net;
 using System.Threading.Tasks;
 using LG.Eloqua.Api.Rest.ClientLibrary.Exceptions;
 using LG.Eloqua.Api.Rest.ClientLibrary.Models.Data.Assets.Campaign;
 using LG.Eloqua.Api.Rest.ClientLibrary.Models.Data.Assets.Email;
+using LG.Eloqua.Api.Rest.ClientLibrary.Models.Dtos;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace LG.Eloqua.Api.Rest.ClientLibrary.Tests.Integration
@@ -57,7 +59,7 @@ namespace LG.Eloqua.Api.Rest.ClientLibrary.Tests.Integration
 
             //Assert
             Assert.IsNotNull(existingContact);
-            Assert.AreEqual(1000,existingContact.Elements.Count);
+            Assert.AreEqual(1000, existingContact.Elements.Count);
             Assert.IsInstanceOfType(existingContact.Elements.First(), typeof(Email));
         }
 
@@ -275,6 +277,28 @@ namespace LG.Eloqua.Api.Rest.ClientLibrary.Tests.Integration
 
             //Assert
             Assert.IsNotNull(result);
+        }
+
+        [TestMethod]
+        public async Task DisableCustomCampaignObjectsAsyncTest()
+        {
+
+            //Arrange
+            var client = new LgEloquaContext(EloquaContext.CreateClient("LeavittGroupAgencyAssociationLLC", Username, Password, new Uri("https://secure.eloqua.com")));
+
+            //Act
+            var result = await client.DisableCustomCampaignObjectsAsync(new List<CustomCampaignObjectDto>
+            {
+                new CustomCampaignObjectDto
+                {
+                    ActivationId = 2110,
+                    InstanceId = 7792181
+                }
+            }
+          );
+
+            //Assert
+            Assert.AreEqual(HttpStatusCode.OK, result.FirstOrDefault()?.Status);
         }
