+++ EloquaClient/Models/Dtos/CustomCampaignObjectDto.cs
@@ -6,6 +6,6 @@ namespace LG.Eloqua.Api.Rest.ClientLibrary.Models.Dtos
     {
         public long InstanceId { get; set; }
         public long ActivationId { get; set; }
-        public HttpStatusCode Status { get; set; }
+        public HttpStatusCode? Status { get; set; }
     }
 }
